The NotFoundException was added and implemented

+++ source/MovieStore.Application/Movies/Handlers/DeleteMovieHandler.cs
@@ -1,6 +1,8 @@
 ﻿using MediatR;
+using MovieStore.Application.Common.Exceptions;
 using MovieStore.Application.Common.Interfaces;
 using MovieStore.Application.Movies.Commands;
+using MovieStore.Domain.Entities;
 using System.Threading;
 using System.Threading.Tasks;
 
@@ -18,6 +20,11 @@ namespace MovieStore.Application.Movies.Handlers
 		public async Task<Unit> Handle(DeleteMovieCommand request, CancellationToken cancellationToken)
 		{
 			var entity = await _context.Movies.FindAsync(request.Id);
+
+			if (entity == null)
+			{
+				throw new NotFoundException(nameof(Movie), request.Id);
+			}
 
 			_context.Movies.Remove(entity);
 			await _context.SaveChangesAsync(cancellationToken);
